# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VerdictCouncil API",
    "summary": "Judicial hearing support system with 9-agent AI analysis pipeline",
    "description": "VerdictCouncil processes judicial cases through a multi-agent pipeline covering case intake, evidence analysis, fact reconstruction, legal knowledge, argument construction, and hearing analysis.\n\n**Authentication:** Cookie-based JWT. Call `POST /api/v1/auth/login` to receive an httpOnly `vc_token` cookie.\n\n**Note:** Swagger UI does not support cookie-based auth for interactive testing. Use a separate HTTP client (curl, httpx) for authenticated requests.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/v1/cases/{case_id}/reopen-request": {
      "post": {
        "tags": [
          "reopen-requests"
        ],
        "summary": "Create reopen request",
        "operationId": "create_reopen_request",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          },
          {
            "name": "vc_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vc Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReopenRequestCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReopenRequestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Case not eligible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Case not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (60 requests/minute per client IP)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds until the rate limit resets",
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/reopen-requests": {
      "get": {
        "tags": [
          "reopen-requests"
        ],
        "summary": "List reopen requests",
        "operationId": "list_reopen_requests",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          },
          {
            "name": "vc_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vc Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReopenRequestListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (60 requests/minute per client IP)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds until the rate limit resets",
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/reopen-requests/{request_id}/review": {
      "patch": {
        "tags": [
          "reopen-requests"
        ],
        "summary": "Review reopen request",
        "operationId": "review_reopen_request",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          },
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          },
          {
            "name": "vc_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vc Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReopenRequestReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReopenRequestResponse"
                }
              }
            }
          },
          "403": {
            "description": "Two-person rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Request not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (60 requests/minute per client IP)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds until the rate limit resets",
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/precedents/search": {
      "post": {
        "tags": [
          "precedent-search"
        ],
        "summary": "Ad-hoc precedent search",
        "description": "Trigger a live PAIR API search for Singapore case law outside the pipeline. Falls back to the curated vector store when PAIR is unavailable. Requires judge role.",
        "operationId": "search_precedents_adhoc",
        "parameters": [
          {
            "name": "vc_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vc Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrecedentSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrecedentSearchResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (60 requests/minute per client IP)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds until the rate limit resets",
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents/{document_id}/excerpt": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get a specific page excerpt from an uploaded document",
        "operationId": "get_document_excerpt",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-indexed page number",
              "title": "Page"
            },
            "description": "1-indexed page number"
          },
          {
            "name": "vc_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vc Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Document Excerpt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (60 requests/minute per client IP)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds until the rate limit resets",
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable error message",
            "examples": [
              "Not found"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response body."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrecedentSearchMetadata": {
        "properties": {
          "source_failed": {
            "type": "boolean",
            "title": "Source Failed",
            "default": false
          },
          "fallback_used": {
            "type": "boolean",
            "title": "Fallback Used",
            "default": false
          },
          "pair_status": {
            "type": "string",
            "title": "Pair Status",
            "default": "ok"
          }
        },
        "type": "object",
        "title": "PrecedentSearchMetadata"
      },
      "PrecedentSearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Query",
            "description": "Free-text search query"
          },
          "jurisdiction": {
            "type": "string",
            "title": "Jurisdiction",
            "description": "Legal domain / jurisdiction",
            "default": "small_claims"
          },
          "max_results": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Results",
            "description": "Maximum results to return",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "PrecedentSearchRequest"
      },
      "PrecedentSearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/PrecedentSearchResultItem"
            },
            "type": "array",
            "title": "Results"
          },
          "metadata": {
            "$ref": "#/components/schemas/PrecedentSearchMetadata"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "searched_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Searched At"
          }
        },
        "type": "object",
        "required": [
          "results",
          "metadata",
          "total"
        ],
        "title": "PrecedentSearchResponse"
      },
      "PrecedentSearchResultItem": {
        "properties": {
          "citation": {
            "type": "string",
            "title": "Citation"
          },
          "court": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Court"
          },
          "outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "reasoning_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Summary"
          },
          "similarity_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Similarity Score"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "citation"
        ],
        "title": "PrecedentSearchResultItem"
      },
...
      "ReopenRequestCreateRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 50,
            "title": "Reason"
          },
          "justification": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Justification"
          }
        },
        "type": "object",
        "required": [
          "reason",
          "justification"
        ],
        "title": "ReopenRequestCreateRequest"
      },
      "ReopenRequestListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReopenRequestResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          }
        },
        "type": "object",
        "title": "ReopenRequestListResponse"
      },
      "ReopenRequestResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "case_id": {
            "type": "string",
            "format": "uuid",
            "title": "Case Id"
          },
          "requested_by": {
            "type": "string",
            "format": "uuid",
            "title": "Requested By"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "justification": {
            "type": "string",
            "title": "Justification"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "review_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Notes"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "case_id",
          "requested_by",
          "reason",
          "justification",
          "status",
          "created_at"
        ],
        "title": "ReopenRequestResponse"
      },
      "ReopenRequestReviewRequest": {
        "properties": {
          "approve": {
            "type": "boolean",
            "title": "Approve",
            "description": "Approve or reject the reopen request"
          },
          "review_notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Notes"
          }
        },
        "type": "object",
        "required": [
          "approve"
        ],
        "title": "ReopenRequestReviewRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "vc_token",
        "description": "JWT token set via httpOnly cookie on login. Call POST /api/v1/auth/login to obtain the cookie."
      }
    }
  },
...
  "security": [
    {
      "cookieAuth": []
    }
  ]
}